'修复导出伦廓线，因点坐标类型定位为int，导致图形圆弧倒角失真，现改为double'

+++ src/slic3r/Utils/ReadPointTool.h
@@ -7,12 +7,12 @@
 
 typedef struct ST_LineData
 {
-    int x1, y1, x2, y2;
+    double x1, y1, x2, y2;
 	 ST_LineData() { 
-		 x1 = 0;
-         y1 = 0;
-         x2 = 0;
-         y2 = 0;
+		 x1 = 0.0;
+         y1 = 0.0;
+         x2 = 0.0;
+         y2 = 0.0;
 	 }
 
 } DL_LineData;
